# docs/architecture/identity/TWIN_API.openapi.v1.yaml
...
openapi: 3.1.0
info:
  title: Marpich Identity Digital Twin API
  version: 1.1.0
  description: Twin registry, synchronization, facets, simulation, prediction
servers:
  - url: /api/v1/identity-twins
paths:
...
  /{twin_ref}/sync:
    post:
      operationId: syncTwin
      summary: Manual or mode-selected synchronization
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncTwinRequest"
      x-permission: twin.write
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    SyncTwinRequest:
      type: object
      properties:
        mode:
          type: string
          enum: [realtime, near_realtime, batch, scheduled, manual, incremental, full]
        facets:
          type: array
          items: { type: string }
        full:
          type: boolean
          default: false
        projection:
          type: object
          additionalProperties: true
...